GETP-206 rename: api.controller에서 공통으로 사용하는 클래스에 대한 api.controller.common 모듈 구성
