+++ src/main/java/es/princip/getp/api/controller/auth/SignUpController.java
@@ -1,9 +1,9 @@
 package es.princip.getp.api.controller.auth;
 
-import es.princip.getp.api.controller.ApiResponse;
-import es.princip.getp.api.controller.ApiResponse.ApiSuccessResult;
 import es.princip.getp.api.controller.auth.dto.request.EmailVerificationCodeRequest;
 import es.princip.getp.api.controller.auth.dto.request.SignUpRequest;
+import es.princip.getp.api.controller.common.dto.ApiResponse;
+import es.princip.getp.api.controller.common.dto.ApiResponse.ApiSuccessResult;
 import es.princip.getp.application.auth.service.SignUpService;
 import es.princip.getp.domain.member.model.Email;
 import jakarta.validation.Valid;
