+++ src/main/java/es/princip/getp/api/controller/common/ControllerSupport.java
@@ -1,4 +1,4 @@
-package es.princip.getp.api.controller;
+package es.princip.getp.api.controller.common;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
