+++ src/main/java/es/princip/getp/api/controller/member/query/MyMemberQueryController.java
@@ -1,6 +1,6 @@
 package es.princip.getp.api.controller.member.query;
 
-import es.princip.getp.api.controller.ApiResponse;
+import es.princip.getp.api.controller.common.dto.ApiResponse;
 import es.princip.getp.api.controller.member.query.dto.response.MemberResponse;
 import es.princip.getp.api.security.details.PrincipalDetails;
 import es.princip.getp.domain.member.model.Member;
