+++ src/main/java/es/princip/getp/api/controller/people/query/MyPeopleProfileQueryController.java
@@ -1,7 +1,7 @@
 package es.princip.getp.api.controller.people.query;
 
-import es.princip.getp.api.controller.ApiResponse;
-import es.princip.getp.api.controller.ApiResponse.ApiSuccessResult;
+import es.princip.getp.api.controller.common.dto.ApiResponse;
+import es.princip.getp.api.controller.common.dto.ApiResponse.ApiSuccessResult;
 import es.princip.getp.api.controller.people.query.dto.peopleProfile.DetailPeopleProfileResponse;
 import es.princip.getp.api.security.details.PrincipalDetails;
 import es.princip.getp.domain.people.query.dao.PeopleDao;
