+++ src/main/java/es/princip/getp/api/controller/project/command/ProjectApplicationController.java
@@ -1,7 +1,7 @@
 package es.princip.getp.api.controller.project.command;
 
-import es.princip.getp.api.controller.ApiResponse;
-import es.princip.getp.api.controller.ApiResponse.ApiSuccessResult;
+import es.princip.getp.api.controller.common.dto.ApiResponse;
+import es.princip.getp.api.controller.common.dto.ApiResponse.ApiSuccessResult;
 import es.princip.getp.api.controller.project.command.dto.request.ApplyProjectRequest;
 import es.princip.getp.api.controller.project.command.dto.response.ApplyProjectResponse;
 import es.princip.getp.api.security.details.PrincipalDetails;
