+++ src/main/java/es/princip/getp/api/exception/DefaultExceptionHandler.java
@@ -1,6 +1,6 @@
 package es.princip.getp.api.exception;
 
-import es.princip.getp.api.controller.ApiErrorResponse;
+import es.princip.getp.api.controller.common.dto.ApiErrorResponse;
 import es.princip.getp.common.exception.DefaultErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
