+++ src/main/java/es/princip/getp/api/security/filter/AccessTokenExceptionFilter.java
@@ -1,8 +1,8 @@
 package es.princip.getp.api.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.princip.getp.api.controller.ApiErrorResponse;
-import es.princip.getp.api.controller.ApiErrorResponse.ApiErrorResult;
+import es.princip.getp.api.controller.common.dto.ApiErrorResponse;
+import es.princip.getp.api.controller.common.dto.ApiErrorResponse.ApiErrorResult;
 import es.princip.getp.api.security.exception.JwtTokenException;
 import es.princip.getp.common.exception.ErrorDescription;
 import jakarta.servlet.FilterChain;
